requirements.txt
langchain==0.1.13
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.33
openai==0.27.6
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
# uncomment to use huggingface llms
#huggingface-hub==0.22.2

# uncomment to use instructor embeddings
#InstructorEmbedding==1.0.1
#sentence-transformers==2.2.2